+++ src/utils/__init__.py
@@ -3,6 +3,7 @@
 from src.utils.sessions import get_session_tag
 from src.utils.trade_logger import export_trade_log, aggregate_trade_logs
 from src.utils.model_utils import download_model_if_missing, download_feature_list_if_missing
+from src.utils.env_utils import get_env_float
 
 __all__ = [
     "get_session_tag",
@@ -10,4 +11,5 @@ __all__ = [
     "aggregate_trade_logs",
     "download_model_if_missing",
     "download_feature_list_if_missing",
+    "get_env_float",
 ]